renamed octree_internal.hpp
