+++ domain/include/cstone/domain/assignment_gpu.cuh
@@ -38,7 +38,7 @@
 #include "cstone/domain/domaindecomp_mpi_gpu.cuh"
 #include "cstone/domain/layout.hpp"
 #include "cstone/primitives/primitives_gpu.h"
-#include "cstone/tree/octree_internal.hpp"
+#include "cstone/tree/octree.hpp"
 #include "cstone/tree/update_mpi.hpp"
 #include "cstone/tree/update_mpi_gpu.cuh"
 #include "cstone/sfc/box_mpi.hpp"
