+++ domain/include/cstone/tree/update_mpi.hpp
@@ -35,7 +35,7 @@
 
 #include "cstone/primitives/mpi_wrappers.hpp"
 #include "cstone/tree/csarray.hpp"
-#include "cstone/tree/octree_internal.hpp"
+#include "cstone/tree/octree.hpp"
 
 namespace cstone
 {
