+++ domain/test/integration_mpi/domain_2ranks.cpp
@@ -33,7 +33,7 @@
 #include <gtest/gtest.h>
 
 #include "cstone/domain/domain.hpp"
-#include "cstone/tree/octree_util.hpp"
+#include "cstone/tree/cs_util.hpp"
 
 #include "coord_samples/random.hpp"
