+++ domain/test/integration_mpi/exchange_keys.cpp
@@ -35,7 +35,7 @@
 
 #include "cstone/domain/exchange_keys.hpp"
 #include "cstone/domain/layout.hpp"
-#include "cstone/tree/octree_util.hpp"
+#include "cstone/tree/cs_util.hpp"
 
 using namespace cstone;
