+++ domain/test/unit/focus/octree_focus.cpp
@@ -32,7 +32,7 @@
 #include "gtest/gtest.h"
 
 #include "cstone/focus/octree_focus.hpp"
-#include "cstone/tree/octree_util.hpp"
+#include "cstone/tree/cs_util.hpp"
 
 namespace cstone
 {
