+++ domain/test/unit/focus/source_center.cpp
@@ -32,7 +32,7 @@
 #include "gtest/gtest.h"
 
 #include "cstone/focus/source_center.hpp"
-#include "cstone/tree/octree_util.hpp"
+#include "cstone/tree/cs_util.hpp"
 
 #include "coord_samples/random.hpp"
