+++ domain/test/unit/traversal/collisions.cpp
@@ -35,7 +35,7 @@
 
 #include "cstone/traversal/collisions.hpp"
 #include "cstone/tree/csarray.hpp"
-#include "cstone/tree/octree_util.hpp"
+#include "cstone/tree/cs_util.hpp"
 
 using namespace cstone;
