+++ domain/test/unit/tree/cs_util.cpp
@@ -37,7 +37,7 @@
 
 #include "cstone/sfc/morton.hpp"
 #include "cstone/tree/csarray.hpp"
-#include "cstone/tree/octree_util.hpp"
+#include "cstone/tree/cs_util.hpp"
 
 using namespace cstone;
