+++ domain/test/unit/tree/csarray.cpp
@@ -32,7 +32,7 @@
 #include "gtest/gtest.h"
 
 #include "cstone/tree/csarray.hpp"
-#include "cstone/tree/octree_util.hpp"
+#include "cstone/tree/cs_util.hpp"
 
 #include "coord_samples/random.hpp"
