+++ domain/test/unit/tree/octree.cpp
@@ -35,8 +35,8 @@
 
 #include "gtest/gtest.h"
 
-#include "cstone/tree/octree_internal.hpp"
-#include "cstone/tree/octree_util.hpp"
+#include "cstone/tree/octree.hpp"
+#include "cstone/tree/cs_util.hpp"
 
 using namespace cstone;
