+++ domain/test/unit_cuda/btree.cu
@@ -36,7 +36,7 @@
 #include "gtest/gtest.h"
 
 #include "cstone/tree/btree.cuh"
-#include "cstone/tree/octree_util.hpp"
+#include "cstone/tree/cs_util.hpp"
 
 using namespace cstone;
