+++ domain/test/unit_cuda/octree.cu
@@ -38,7 +38,7 @@
 #include "coord_samples/random.hpp"
 #include "cstone/tree/csarray.hpp"
 #include "cstone/tree/csarray_gpu.h"
-#include "cstone/tree/octree_util.hpp"
+#include "cstone/tree/cs_util.hpp"
 #include "cstone/tree/update_gpu.cuh"
 
 using namespace cstone;
